[BE] feat: jql 수정

+++ FE/ConvenienceSightApp/app/src/main/java/com/example/lookey/data/model/allergy/AllergyDeleteResponse.kt
@@ -1,7 +1,7 @@
 package com.example.lookey.data.model.allergy
 
 data class AllergyDeleteRequest(
-    val allergy_id: Int
+    val allergy_id: Long
 )
 
 data class AllergyDeleteResponse(
