+++ FE/ConvenienceSightApp/app/src/main/java/com/example/lookey/domain/repo/AlllergyRepository.kt
@@ -5,6 +5,6 @@ import com.example.lookey.domain.entity.Allergy
 interface AllergyRepository {
     suspend fun list(): List<Allergy>
     suspend fun search(q: String): List<Allergy>
-    suspend fun add(allergyId: Int)
-    suspend fun delete(allergyId: Int)
+    suspend fun add(allergyId: Long)
+    suspend fun delete(allergyId: Long)
 }
